Move Fortran interface to a separate file
